refactor(web): extract RouteFormModal; edit routes in place from detail page
